soc: qcom: peripheral-loader: Fix uninitialized structure
Fortunately, this is a false positive. This variable is initialized
within an if statement on line 179 and used again without that same
conditional on line 220. However, the for loop on line 219 will not
run if the if statement was not hit because ss_mdump_seg_cnt_pdr is
initialized to 0 so there is no way for this variable to be used
uninitialized. Regardless, GCC still warns about it so initialize it
to NULL.

Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>
Signed-off-by: Adam W. Willis <return.of.octobot@gmail.com>