Reapply ecd300b ("qcacld: defer hdd initialization and rely on userspace writing to /dev/wlan")
Wi-Fi HAL writes "ON" or "OFF" to /dev/wlan.

Use this method to initialize hdd as it's a safer way to ensure both
wlan_mac.bin and WCNSS_qcom_cfg.ini is ready to be read.

This also eliminates the needs for horrible hack to read userspace file.

Adapted for LA.UM.7.1.r1-14500-sm8150.0 (5.2.03.9H)

Signed-off-by: Park Ju Hyung <qkrwngud825@gmail.com>
Signed-off-by: Adam W. Willis <return.of.octobot@gmail.com>