psi: clarify the Kconfig text for the default-disable option
The current help text caused some confusion in online forums about whether
or not to default-enable or default-disable psi in vendor kernels.  This
is because it doesn't communicate the reason for why we made this setting
configurable in the first place: that the overhead is non-zero in an
artificial scheduler stress test.

Since this isn't representative of real workloads, and the effect was not
measurable in scheduler-heavy real world applications such as the
webservers and memcache installations at Facebook, it's fair to point out
that this is a pretty cautious option to select.

Link: http://lkml.kernel.org/r/20190129233617.16767-1-hannes@cmpxchg.org
Signed-off-by: Johannes Weiner <hannes@cmpxchg.org>
Reviewed-by: Andrew Morton <akpm@linux-foundation.org>
Cc: Mel Gorman <mgorman@techsingularity.net>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>
Signed-off-by: Stephen Rothwell <sfr@canb.auug.org.au>

(in linux-next: https://git.kernel.org/pub/scm/linux/kernel/git/next/linux-next.git/commit/?id=8ed912a5006bb017c63ba343ffc77e7b993634b4)

Bug: 111308141
Test: modified lmkd to use PSI and tested using lmkd_unit_test

Change-Id: I34bc09dcab5304b2e467970b709b8b31036a5481
Signed-off-by: Suren Baghdasaryan <surenb@google.com>
Signed-off-by: Adam W. Willis <return.of.octobot@gmail.com>