Revert "clk: qcom: clk-cpu-osm: Limit performance cluster to 2323200 kHz"
I'm not convinced of the viability of this, the supporting evidence of
it's efficacy is a bit misleading in that it compares the power
consumption and performance of 2323 MHz against 2649 MHz, and not the
maximum frequency of 2803 MHz. While this suggests the potetial for more
substantial power savings, it also obscures the true performance loss
associated with this underclock. Perf cluster only utilizes max
frequency approximately 3-4% of the time during a cycle, thus I do not
believe that the benefits of mandating this cap outweigh the drawbacks.

This reverts commit 69aac76.

Signed-off-by: Adam W. Willis <return.of.octobot@gmail.com>