neutrino: Tentatively reduce MAX_STUNE_BOOST
Although I have yet to see this occur during practical usage, there is
evidence to suggest that there is a bug present in CAF's SDM845
scheduler code which can cause sched boosted top-app tasks to run on LP
cluster with a boost margin calculated for the perf cluster. This has
the potential to cause LP frequencies to be locked at 1766 MHz
indefinitely. For the time being, we will lower max boost schedtune to
match input boost's value of 15 which has been shown to mitigate this
behavior.

Signed-off-by: Adam W. Willis <return.of.octobot@gmail.com>