neutrino: Enable ZRAM/ZPOOL
It appears as though OnePlus has enabled swap in the latest OxygenOS Betas. While I
don't necessarily agree with the rationale, we will give this strategy the benefit of
the doubt and perform some preliminary testing with a small, low swap zRAM cache. If
nothing else, we will preemptively have the code in place and if this proves to be
snakeoil it can be reassessed in the future.

Subsequent to the following commits:
dfbae5c ("zram: Add compile-time support for choosing default allocator")
1a14e5c ("zram: allow choosing default compressor via config")

Signed-off-by: Adam W. Willis <return.of.octobot@gmail.com>