FIXUP: kernel: Boost to the max for a short amount of time when zygote forks
Synchronize with kerneltoast/android_kernel_google_wahoo@52b1556

In order to boost during the period from when a new app is forked to the
time that it is added to the top-app cpuset (which is a signal to
subsequently boost again), boost for a short amount of time when zygote
forks. Not all processes forked from zygote are top apps, hence the
decoupled boosting setup.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
[@0ctobot: This amends commit e838d8b ("kernel: Boost to the max for a short amount of time when zygote forks")]
Signed-off-by: Adam W. Willis <return.of.octobot@gmail.com>