Revert "kbuild: Support Polyhedral Loop Optimizations for LLVM"
There's no practical reason that this needs to be called as a Kconfig
option when these flags can just as easily be moved to our build script.
Additionally, this only serves to create an additional step and
potential headache for those attempting to compile the kernel inline.
Finally, this interferes with compilation when CONFIG_LTO_CLANG is
enabled. Just get rid of it.

This reverts commit 342527d.

Signed-off-by: Adam W. Willis <return.of.octobot@gmail.com>