Remove support for clang SafeStack
This is a deprecated feature that Google themselves have opted to
drop due to potential instability and memory corruption issues. We
were never going to use it anyway, so there's no reason for all of
this extraneous code to be hanging around.

This reverts the following commits:
2382236 ("soc: qcom: pil: disable SafeStack for pil_assign_mem_*")
1f4f33f ("arm64: add support for clang SafeStack")
a54f9c5 ("add support for clang SafeStack")

Signed-off-by: Adam W. Willis <return.of.octobot@gmail.com>