Revert "drm: Enable bridges asynchronously as early as possible"
Apparently, this commit exacerbates the static/CRT screen
flickering effect experienced by some users with varying frequency
when powering on the display. I'm not convinced that this commit is
solely responsible for the behavior as it reportedly can occur on
the stock OOS kernel as well. However, we'll try rolling this back
in hopes attenuating the annoyance.

This reverts the following commits:
51acb77 ("FIXUP: drm: Enable bridges asynchronously as early as possible")
97682fa ("drm: Enable bridges asynchronously as early as possible")

Suggested-by: acuicultor <ramoncuenca4@gmail.com>
Signed-off-by: Adam W. Willis <return.of.octobot@gmail.com>