[TEMP] initramfs: add skip_override parameter
We will be temporarily relying on this skip_override hack as it has
been reported that the current best practice of using magiskboot to
hexpatch the kernel at install time has proven unreliable in
conjunction with Magisk 19.4 on initial Android 10 builds for this
device. I'm currently unsure if this is a Magisk issue, a software
issue, or some combination of the two. Until I can get hands on
with Q myself and explore this further, we'll fall back to this
approach to root preservation as other kernels still using it do
not appear to be experiencing these issues.

Signed-off-by: Adam W. Willis <return.of.octobot@gmail.com>