neutrino: Disable PROJECT_INFO by default
Subsequent to 0057ddf327f2065b1cee338a7cef5a95b05bff03 ("soc: qcom: Restore OnePlus project_info framework"),
this ultimately did not reconcile the issue that I was trying to address.
Despite restoring a presumably functional project_info driver, McLaren
specific cosmetics are still failing to be enabled on OxygenOS. At this
point I have no idea why this is the case so we can continue acting as
though project_info doesn't exist as it currently isn't doing us any
favors. However, I will keep the driver intact but disabled for the time
being in hopes that I may become intimately familiar with how the McLaren
check is performed in the future and be able to properly ascertain what
exactly is going on here. Whatever the case may be, I have to assume that
project_info will be required in some capacity.

Signed-off-by: Adam W. Willis <return.of.octobot@gmail.com>