[TEMP] writeback: Restore stock dirty ratios
We'll be dropping the NeutrinoSettings module in an effort to sever
dependancies on Magisk. The majority of our critical configuration
is already handled by the kernel, essentially memory settings are
the only remaining parameters which userspace has a tendency to
override. Given their limited importance to overall UX, rollback
dirty ratio values to match stock defaults and let userspace decide
what it wants until we can implement proper init protection for
these variables.

Signed-off-by: Adam W. Willis <return.of.octobot@gmail.com>