qcacld-3.0/qcacmn: Reset to 'LA.UM.7.3.r1-08200-sdm845.0' merge
CAF's sdm845 Android 10 tag now uses the 5.2.03 qcacld-3.0 release.
This is the same base used for the sm8150 WLAN stack, which we have
already attempted to implement with unsatisfactory results in the
past. Unfortunately, the same seems to hold true for what are now
the 845 drivers resulting in renewed and vigorous user complaints
of various connectivity issues, with a particular focus on
significant speed defecits for 5GHz connections.

I am getting so incredibly tired of shuffling through WLAN drivers,
but nonetheless, rollback to the most recent 845 Pie tag for the
time being.

Signed-off-by: Adam W. Willis <return.of.octobot@gmail.com>